requirements.txt
streamlit
sentence-transformers
faiss-cpu
PyPDF2
groq
numpy